% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova2.R
\name{Anova2}
\alias{Anova2}
\title{Anova Tables for Various Statistical Models}
\usage{
Anova2(mod, ...)
}
\description{
This function is a modification of the car::Anova function, testing the
significativity of main effects without the interactions terms.
}
\author{
Marc Choisy
}
