% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Boostrap using given data and statistic}
\usage{
bootstrap(fun, data, nreps)
}
\arguments{
\item{fun}{function to calculate on each sample}

\item{data}{data to use for boostrapping. Should be a respresentative sample}

\item{nreps}{number of times to bootstrap}
}
\description{
Boostrap using given data and statistic
}
\examples{
x <- rnorm(100)
bootstrap(mean, x, 1000)
}
