% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_interactions.R
\name{drop_interactions}
\alias{drop_interactions}
\title{Drop Interactions}
\usage{
drop_interactions(x)
}
\arguments{
\item{formula}{an object of class formula}
}
\description{
This function drop interaction terms from a statistical model formula
}
\details{
Returns a formula without interaction terms
}
\author{
Marc Choisy
}
