% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{labelBootResults}
\alias{labelBootResults}
\title{Label Bootstrapped Results}
\usage{
labelBootResults(results, lBound, uBound)
}
\arguments{
\item{results}{a vector, data from bootstrapping}

\item{lBound}{lower bound of confidence interval}

\item{uBound}{upper bound of confidence interval}
}
\value{
vector of labels corresponding to result values
}
\description{
labels bootstrapped results. We use this to create colored histograms.
}
\examples{
x <- rnorm(100)
labelBootResults(x, -1, 1)
}
