% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisHypothesisTests.R
\name{labelDisPDF}
\alias{labelDisPDF}
\title{Label discrete PDF}
\usage{
labelDisPDF(x, obsVal, expVal)
}
\arguments{
\item{x}{x value}

\item{obsVal}{observed event}

\item{expVal}{expected value}
}
\value{
vector of labels for x value in relation to observed event
}
\description{
labels a discrete pdf
}
\examples{
labelDisPDF(0:10, 3, 5)
}
