% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{mcDChiSq}
\alias{mcDChiSq}
\title{Density of Chi-Square distribution}
\usage{
mcDChiSq(x, degFree, ...)
}
\arguments{
\item{x}{x value}

\item{degFree}{degrees of freedom}

\item{...}{optional additional parameters which are ignored}
}
\value{
density of given Chi-Square dist. at x
}
\description{
Density of Chi-Square distribution
}
