% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{mcDF}
\alias{mcDF}
\title{Density of F-distribution}
\usage{
mcDF(x, degFree1, degFree2, ...)
}
\arguments{
\item{x}{x value}

\item{degFree1}{degrees of freedom 1}

\item{degFree2}{degrees of freedom 2}

\item{...}{optional additional parameters which are ignored}
}
\value{
density of given F-dist. at x
}
\description{
Density of F-distribution
}
\examples{
mcDF(1,5,5)

}
