% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{mcDNorm}
\alias{mcDNorm}
\title{dnorm but with more arguments}
\usage{
mcDNorm(x, mean = 0, sd = 1, log = FALSE, ...)
}
\arguments{
\item{x}{x value}

\item{mean}{mean of normal distribution}

\item{sd}{std. dev. of noraml distribution}

\item{log}{logical; if TRUE probabilities are given as log(p). See stats::dnorm}

\item{...}{extra parameters which are ignored}
}
\value{
density of normal distribution
}
\description{
compute density of normal distribution while allowing for more arguments which are ignored
}
\examples{
mcDNorm(x = 0)

}
