% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{mcDT}
\alias{mcDT}
\title{Density of t-distribution}
\usage{
mcDT(x, degFree, ...)
}
\arguments{
\item{x}{x value}

\item{degFree}{degrees of freedom}

\item{...}{optional additional parameters which are ignored}
}
\value{
density of given t-dist. at x
}
\description{
Density of t-distribution
}
\examples{
mcDT(0,1)

}
