% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstats.R
\docType{package}
\name{mcstats}
\alias{mcstats}
\alias{mcstats-package}
\title{mcstats: A package for complementing the base stats package.}
\description{
mcstats: A package for complementing the base stats package.
}
