% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{mosaicplot}
\alias{mosaicplot}
\title{Mosaic Plot}
\usage{
mosaicplot(x)
}
\arguments{
\item{x}{must be a matrix with each row and column labelled}
}
\value{
mosaic plot showing observed proportions, colored by residuals from chi-sq. test
}
\description{
Mosaic Plot
}
\examples{
x <- matrix(runif(9,5,100), ncol = 3, dimnames = list(c("Yes1", "No1", "Maybe1"),
c("Yes2", "No2", "Maybe2")))
mosaicplot(x)
}
