% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTests.R
\name{normPDF}
\alias{normPDF}
\title{Normal Distribution PDF}
\usage{
normPDF(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{x value}

\item{mu}{mean of distribution}

\item{sigma}{std. dev. of distribution}
}
\description{
Returns density of t distribution (uses the normal distribution for now) if in shaded region
}
\examples{
normPDF(0)
}
