% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Estimate proportions with confidence intervals}
\usage{
proportion(x, n, ci = 0.95)
}
\arguments{
\item{x}{number of successes, or a vector of length 2 giving the numbers of
successes and failures, respectively.}

\item{n}{number of trials; ignored if x has length 2.}

\item{ci}{confidence level for the returned confidence interval.}
}
\description{
This is a wrapper around the \code{\link[stats]{binom.test}} function.
}
\seealso{
\code{\link[stats]{binom.test}}
}
\author{
Marc Choisy
}
