% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range2.R
\name{range2}
\alias{range2}
\title{Range of values}
\usage{
range2(..., n = 1, na.rm = FALSE)
}
\arguments{
\item{...}{Any numeric or character objects.}

\item{n}{Positive numeric single value indicating the factor we wish to
reduce (n < 1) or extand (x > 1) the range.}

\item{na.rm}{Logical indicating if NAs should be omitted.}
}
\value{
Same output as the base R \code{range()} function.
}
\description{
Wrapper around the base R \code{range()} function that allows to expand the
width of the range.
}
\details{
The range is reduced or extended symetrically, both sides of the current
range.
}
\examples{
x <- rnorm(100)
range(x)
range2(x, n = 2)
diff(range2(x, n = 2)) / 2
diff(range2(x, n = .5)) / .5

}
