% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTests.R
\name{shadePDF}
\alias{shadePDF}
\title{Used to shade in a PDF}
\usage{
shadePDF(x, fun, testStat, ...)
}
\arguments{
\item{x}{x value}

\item{fun}{density function to use}

\item{testStat}{test statistic value}

\item{...}{optional parameters passed to density function}
}
\value{
density if outside of extreme event region
}
\description{
Returns density with extreme event region having NAs
}
