% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{showANOVA}
\alias{showANOVA}
\title{Show results of ANOVA}
\usage{
showANOVA(formula, data = NULL, verbose = 1, ...)
}
\arguments{
\item{formula}{formula specifying a model.}

\item{data}{data on which to perform ANOVA}

\item{verbose}{if verbose > 0 the resulting graph is printed}

\item{...}{Arguments passed to lm. See \link[stats]{aov} for more detail}
}
\value{
output of call to \link[stats]{aov}
}
\description{
Visualization of distributional results of ANOVA. Please see \link[stats]{aov} for more
information on parameters
}
\examples{
showANOVA(yield ~  N + P + K, npk)
}
