% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{showOLS}
\alias{showOLS}
\title{Show hypothesis tests from OLS}
\usage{
showOLS(formula, data, verbose = 1)
}
\arguments{
\item{formula}{forumula for regression. Passed to \link[stats]{lm}}

\item{data}{data for regression. Passed to \link[stats]{lm}}

\item{verbose}{if verbose > 0 the resulting graph is printed}
}
\value{
model object resulting from the regression
}
\description{
Show hypothesis tests from OLS
}
\examples{
showOLS(mpg ~ cyl + disp, mtcars)
}
