% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisHypothesisTests.R
\name{showProp.Test}
\alias{showProp.Test}
\title{Show results of proportion test using \link[stats]{binom.test}}
\usage{
showProp.Test(x, n, p = 0.5)
}
\arguments{
\item{x}{x value}

\item{n}{number of repetitions}

\item{p}{probability of success in one Bernoulli trial}
}
\value{
output of call to \link[stats]{binom.test}
}
\description{
Show results of proportion test using \link[stats]{binom.test}
}
\examples{
showProp.Test(3, 10)
}
