% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTests.R
\name{showT.Test}
\alias{showT.Test}
\title{Conduct t-test}
\usage{
showT.Test(group1, group2 = NULL, mu = 0, verbose = 1)
}
\arguments{
\item{group1}{continuous data to test}

\item{group2}{optional: second group to include for two sample t-test}

\item{mu}{optional: mean to test against for one-sample t-test}

\item{verbose}{default is 1 which will create a graph. To turn this off use verbose = 0.}
}
\value{
results of call to t.test
}
\description{
Runs t-test and outputs graph for interpretation
}
\examples{
x <- rnorm(100)
showT.Test(x, verbose = 0)
showT.Test(x)

}
