% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisHypothesisTests.R
\name{showWilcoxon.Test}
\alias{showWilcoxon.Test}
\title{Visualize Results of a Wilcoxon Rank Sum Test}
\usage{
showWilcoxon.Test(cat1, cat2)
}
\arguments{
\item{cat1}{vector of measurements for group 1}

\item{cat2}{vector of measurements for group 2}
}
\value{
list containing p-value and test statistic
}
\description{
Visualize Results of a Wilcoxon Rank Sum Test
}
