% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTests.R
\name{showXtremeEvents}
\alias{showXtremeEvents}
\title{Highlight extreme events}
\usage{
showXtremeEvents(testStat, densFun, ...)
}
\arguments{
\item{testStat}{test statistic}

\item{densFun}{function that computes appropriate density}

\item{...}{extra arguments passed to density function}
}
\value{
results of call testFun
}
\description{
Make graph highlighting events more extreme than observed sample
}
\examples{
x <- rnorm(100)
showT.Test(x, verbose = 0)
showT.Test(x)

}
