% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsegment_center.R
\name{subsegment_center}
\alias{subsegment_center}
\title{Centers of subsegments}
\usage{
subsegment_center(x, n)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{n}{A vector on integer values specifying how many subsegments we wish
per initial segment. It should either contain one sigle value or
have the same length as x.}
}
\value{
A list of vectors of numeric values. Each slot of this list
corresponds to a segments of x and contains the centers fo the subsegments.
}
\description{
Takes a vector of numeric values and return a vector of numeric values that
correspond to the centers of n subsegments.
}
\details{
The vector x = c(x1, x2, x3, x4, ...) is interpreted as the centers of
segments [y1, y2], [y2, y3], [y3, y4], etc... The function splits each of
these segments into a number of subsegments of equal size and return the
center of these segments.
}
\examples{
subsegment_center(1:3, 5)

}
