% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTests.R
\name{tShade}
\alias{tShade}
\title{T-Distribution PDF}
\usage{
tShade(x, lBound, uBound, mu = 0, sigma = 1)
}
\arguments{
\item{x}{x value}

\item{lBound}{x value that should be shaded below}

\item{uBound}{x value that should be shaded above}

\item{mu}{mean of distribution}

\item{sigma}{std. dev. of distribution}
}
\description{
Returns density of t distribution (uses the normal distribution for now) if in shaded region
}
