% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tTest}
\alias{tTest}
\title{Runs t-test and outputs graph for interpretation}
\usage{
tTest(group1, group2 = NULL, mu = 0)
}
\arguments{
\item{group1}{continuous data to test}

\item{group2}{optional: second group to include for two sample t-test}

\item{mu}{optional: mean to test against for one-sample t-test}
}
\description{
Runs t-test and outputs graph for interpretation
}
\examples{
x <- rnorm(100)
tTest(x)

}
