% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtsHypothesisTests.R
\name{showMcNemarTest}
\alias{showMcNemarTest}
\title{Visualize results of McNemar's Test}
\usage{
showMcNemarTest(x, y = NULL, correct = TRUE, verbose = 1)
}
\arguments{
\item{x}{two dimensional contingency table as a matrix or a factor object}

\item{y}{factor object, ignored if x is a matrix}

\item{correct}{logical indicating whether or not to perform continuity correction}

\item{verbose}{if verbose > 0 the resulting graph is printed}
}
\value{
results of call to \link[stats]{mcnemar.test}
}
\description{
relevant parameters are passed to \link[stats]{mcnemar.test}
}
\examples{
Performance <-
matrix(c(794, 86, 150, 570),
    nrow = 2,
    dimnames = list("1st Survey" = c("Approve", "Disapprove"),
                     "2nd Survey" = c("Approve", "Disapprove")))
showMcNemarTest(Performance)

}
