% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisHypothesisTests.R
\name{showXtremeEventsDis}
\alias{showXtremeEventsDis}
\title{Show Extreme Events from a Discrete Distribution}
\usage{
showXtremeEventsDis(testID, obsVal, expVal, xVals, probFun, ...)
}
\arguments{
\item{testID}{name of test being performed. This is used to title the graph}

\item{obsVal}{observed x value}

\item{expVal}{expected x value}

\item{xVals}{domain of x (possible values)}

\item{probFun}{probability mass function for the given distribution}

\item{...}{addition arguments passed to probFun}
}
\value{
graph coloring events by how extreme they are under the null hypothesis
}
\description{
Show Extreme Events from a Discrete Distribution
}
\examples{
showXtremeEventsDis("Prop. Test", 3, 5, 0:10, probFun = dbinom, size = 10, prob = 0.5)
}
